QtCoap module
This is a new module allowing the use of the CoAP protocol with Qt.
Implemented features :
- Send GET/POST/PUT/DELETE requests
- Discover resources
- Observe resources and cancel the observation
- Blockwise requests and replies
Requests (and replies) can be confirmable or non-confirmable.
Some options can be added to the request.
Replies can be received in a separate or piggybacked message.
Lost confirmable messages are retransmitted.

Work in progress.

Change-Id: I31e0e20a4f19bdc6d6489281fde73a4ff848eda4